Move table selection functions into a Trait
